Moved get_ruptures in commonlib.calc
